Add a new “getEntityStatus” utility function
