+++ services/ui-src/src/components/tables/EntityStatusIcon.tsx
@@ -7,28 +7,13 @@ import unfinishedIcon from "assets/icons/icon_error_circle_bright.png";
 import successIcon from "assets/icons/icon_check_circle.png";
 import { useContext, useMemo } from "react";
 import { ReportContext } from "components/reports/ReportProvider";
-import { mapValidationTypesToSchema } from "utils";
-import { object } from "yup";
+import { getMlrEntityStatus } from "utils/tables";
 
 export const EntityStatusIcon = ({ entity }: Props) => {
   const { report } = useContext(ReportContext);
 
   const entityComplete = useMemo(() => {
-    const reportFormValidation = Object.fromEntries(
-      Object.entries(report?.formTemplate.validationJson ?? {}).filter(
-        ([key]) => key.includes("report_") || key.includes("state_")
-      )
-    );
-
-    const formValidationSchema =
-      mapValidationTypesToSchema(reportFormValidation);
-    const formResolverSchema = object(formValidationSchema || {});
-
-    try {
-      return formResolverSchema.validateSync(entity);
-    } catch (err) {
-      return false;
-    }
+    return report ? getMlrEntityStatus(report, entity) : false;
   }, [report]);
 
   return (
